[Austin] reorganize tests
