+++ client/src/Forms/__tests__/FormComponent.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import renderer from "react-test-renderer";
-import Form from './FormComponent';
+import Form from '../FormComponent';
 
 test('termRequest?', () => {
   const component = renderer.create(<Form title="Request a Term" inputs={[{labelInput:'What term do you want defined?', isRequired:true}]} />);
